新增删除批量题库、题目

+++ src/main/java/com/zzuli/service/RecordService.java
@@ -65,5 +65,5 @@ public interface RecordService extends IService<Record> {
      * @param userId
      * @return
      */
-    Integer getWrongCount(Long userId);
+    Long getWrongCount(Long userId);
 }
